move search thread to it's own file
